1.18_experimental-snapshot-1

+++ minecraft/src/net/minecraft/world/level/levelgen/structure/templatesystem/AxisAlignedLinearPosTest.java
@@ -44,7 +44,7 @@ public class AxisAlignedLinearPosTest extends PosRuleTest {
 		float h = (float)Math.abs((blockPos2.getZ() - blockPos3.getZ()) * direction.getStepZ());
 		int i = (int)(f + g + h);
 		float j = random.nextFloat();
-		return (double)j <= Mth.clampedLerp((double)this.minChance, (double)this.maxChance, Mth.inverseLerp((double)i, (double)this.minDist, (double)this.maxDist));
+		return j <= Mth.clampedLerp(this.minChance, this.maxChance, Mth.inverseLerp((float)i, (float)this.minDist, (float)this.maxDist));
 	}
 
 	@Override
